ref: merge main and chat interfaces + remove local installation scripts
