+++ helper.py
@@ -9,7 +9,7 @@ def jsonmaker():
     # Set default values for variables
     api_key = ""
     endpoint = "https://api.openai.com/v1/"
-    install_flag = 0
+    ollama_flag = 0
     # Check whether the userinfo.json file exists
     if os.path.exists("userinfo.json"):
         # Check whether the API key is set
@@ -18,38 +18,38 @@ def jsonmaker():
             if "api_key" in userinfo:
                 api_key = userinfo["api_key"]
                 endpoint = "https://api.openai.com/v1/"
-                if "install_flag" not in userinfo:
-                    install_flag = 0
+                if "ollama_flag" not in userinfo:
+                    ollama_flag = 0
             else:
                 api_key = ""
                 endpoint = "http://localhost:8080/v1"
-                if "install_flag" not in userinfo:
-                    install_flag = 0
+                if "ollama_flag" not in userinfo:
+                    ollama_flag = 0
 
 
     # If the userinfo.json file does not exist, create it
     else:
         api_key = ""
         endpoint = "https://api.openai.com/v1/"
-        install_flag = 0
+        ollama_flag = 0
 
         userinfo = {
             "api_key": api_key,
             "endpoint": endpoint,
-            "install_flag": install_flag
+            "ollama_flag": ollama_flag
         }
         with open("userinfo.json", "w") as f:
             json.dump(userinfo, f, indent=4)
 
 
-    return api_key, endpoint, install_flag
+    return api_key, endpoint, ollama_flag
 
 
-api_key, endpoint, install_flag = jsonmaker()    # Calling for the other files accessing the variables in this file.
+api_key, endpoint, ollama_flag = jsonmaker()    # Calling for the other files accessing the variables in this file.
 
 # ------------------- LICENSE -------------------
 # Docuchat, a smart knowledge assistant for your documents.
-# Copyright © 2024 xTellarin
+# Copyright © 2025 alxc75
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
